pyproject.toml
[tool.poetry]
name = "fastapi-demo"
version = "0.1.0"
description = ""
authors = ["codex <gwonhyeong.tak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
httpx = "^0.24.1"
anyio = "^3.7.1"
pytest = "^7.4.0"
trio = "^0.22.2"
dependency-injector = "^4.41.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.2"
sqlalchemy = "^2.0.19"
asyncpg = "^0.28.0"
greenlet = "^2.0.2"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
